[RW-2271][RW-2297][risk=no] concept set details react (#2118)
