--- ui/src/app/views/concept-set-details/component.css
@@ -1,116 +1,0 @@
-.concept-set-navigation * {
-  display: block;
-}
-
-.concept-set-header {
-  display: flex;
-}
-
-.concept-set-details {
-  /* Expand the details to take as much of the header as possible. */
-  flex-grow: 1;
-  padding-bottom: 1.5rem;
-  padding-right: 10%;
-}
-
-.action-menu-dropdown {
-  /* The edit icon size. */
-  height: 24px;
-  margin-left: .5rem;
-  margin-top: .5rem;
-}
-
-.action-item {
-  text-align: left;
-  color: #4A4A4A;
-  background-color: white;
-  fill: none;
-  cursor: pointer;
-  border: none;
-}
-
-.action-item:hover {
-  background-color: #E0EAF1;
-}
-
-.ellipsis-vertical {
-  fill: #2691D0;
-}
-
-.concept-set-name * {
-  color: #2F2E7E;
-  display: inline-block;
-}
-
-.concept-set-name h2 {
-  font-size: 20px;
-  font-weight: 600;
-  margin-top: 0;
-}
-
-.edit-name {
-  margin-bottom: .8rem;
-  /* To match .btn styling. */
-  margin-top: 6px;
-}
-
-.edit-button {
-  border: 0;
-  margin-top: .2rem;
-}
-
-.edit-button:hover {
-  border: 1px solid;
-}
-
-.concept-set-details p {
-  color: rgba(0,0,0,0.87);
-  margin-bottom: 1.5rem;
-  margin-top: 0;
-}
-
-.edit-description,
-.edit-buttons {
-  margin-bottom: .5rem;
-}
-
-.concept-set-tags {
-  margin-bottom: .5rem;
-}
-
-.concept-set-tags * {
-  color: rgba(0,0,0,0.87);
-  font-weight: 600;
-  margin-right: 2rem;
-  white-space: nowrap;
-}
-
-.nav-button {
-  background-color: #FFFFFF;
-  border-color: #2691D0;
-  color: #2691D0;
-  text-align: left;
-  text-transform: none;
-}
-
-.concept-set-navigation * {
-  width: 10rem;
-}
-
-.nav-button .icon {
-  display: inline-block;
-  height: .8rem;
-  width: .8rem;
-}
-
-.remove-icon {
-  fill: white;
-  height: 1.2rem;
-  width: 1.2rem;
-  margin-right: .25rem;
-}
-
-.concept-table-wrap {
-  margin-left: 1rem;
-  margin-right: 1rem;
-}
