--- ui/src/app/views/concept-set-details/component.html
@@ -1,101 +1,0 @@
-<app-top-box *ngIf="conceptSet">
-  <div class="concept-set-header">
-    <clr-dropdown class="action-menu-dropdown">
-      <!-- TODO: Once clone is added to the menu, never disable the dropdown. -->
-      <button type="button" class="resource-menu" [disabled]="!canEdit" clrDropdownTrigger>
-        <clr-icon shape="ellipsis-vertical" class="ellipsis-vertical"></clr-icon>
-      </button>
-      <clr-dropdown-menu *clrIfOpen>
-        <button clrDropdownItem class="action-item action-edit"
-            (click)="editing=true"
-            [class.disabled]="!canEdit"
-            [disabled]="!canEdit">
-          <clr-icon shape="pencil"></clr-icon>
-          Edit
-        </button>
-        <button clrDropdownItem class="action-item action-delete"
-            (click)="confirmDeleting=true"
-            [class.disabled]="!canEdit"
-            [disabled]="!canEdit">
-          <clr-icon shape="trash"></clr-icon>
-          Delete
-        </button>
-      </clr-dropdown-menu>
-    </clr-dropdown>
-    <div class="concept-set-details">
-      <div class="concept-set-name">
-        <h2 *ngIf="!editing">{{conceptSet.name}}</h2>
-        <input *ngIf="editing" class="edit-name" type="text" autofocus [(ngModel)]="editName">
-        <app-edit-icon
-            class="edit-icon"
-            (click)="editing = true"
-            [disabled]="!canEdit"
-            [style]="{marginTop: '0.1rem'}"></app-edit-icon>
-      </div>
-      <p *ngIf="!editing">{{conceptSet.description}}</p>
-      <textarea *ngIf="editing" class="edit-description" type="text" [(ngModel)]="editDescription"></textarea>
-      <div *ngIf="editing" class="edit-buttons">
-        <span role="tooltip" aria-haspopup="true" class="tooltip tooltip-sm tooltip-top-left">
-          <button type="button" class="btn btn-primary submit-edit-button"
-              (click)="submitEdits()" [clrLoading]="editSubmitting" [disabled]="!validateEdits()">
-            Save
-          </button>
-          <span *ngIf="!editName" class="tooltip-content">A name is required</span>
-        </span>
-        <button type="button" class="btn btn-outline cancel-edit-button" (click)="editing=false">Cancel</button>
-      </div>
-      <div class="concept-set-tags">
-        <span>Participant count: {{conceptSet.participantCount}}</span>
-        <span>Domain: {{conceptSet.domain | titlecase}}</span>
-      </div>
-      <!-- The empty state. -->
-      <button *ngIf="!conceptSet.concepts?.length" class="btn nav-button"
-              routerLink="/workspaces/{{wsNamespace}}/{{wsId}}/concepts"
-              [queryParams]="{domain: conceptSet.domain}" routerLinkActive="active">
-        <clr-icon shape="search" class="icon search-icon"></clr-icon>
-        Add concepts to set
-      </button>
-    </div>
-    <div class="concept-set-navigation" style="display: inline-block;">
-      <button class="btn nav-button" routerLinkActive="active"
-          routerLink="/workspaces/{{wsNamespace}}/{{wsId}}/concepts"
-          [queryParams]="{domain: conceptSet.domain}">
-        <clr-icon shape="search" class="icon search-icon"></clr-icon>
-        Add concepts to set
-      </button>
-      <button class="btn nav-button" routerLinkActive="active"
-          routerLink="/workspaces/{{wsNamespace}}/{{wsId}}/concepts/sets">
-        <clr-icon shape="grid-view" class="icon search-icon"></clr-icon>
-        Return to concept sets
-      </button>
-    </div>
-  </div>
-  <div *ngIf="conceptSet.concepts?.length" class="concept-table-wrap">
-    <app-concept-table [concepts]="conceptSet.concepts" [placeholderValue]="'No Concepts Found'"
-    [onSelectConcepts]="onSelectConcepts" [selectedConcepts]="selectedConcepts"></app-concept-table>
-  </div>
-  <app-sliding-fab [disable]="disableRemoveFab" [expanded]="'Remove from set'"
-                   [submitFunction]="removeFab" [iconShape]="'trash'">
-  </app-sliding-fab>
-</app-top-box>
-
-<app-confirm-delete-modal *ngIf="confirmDeleting"
-                          [resourceType]="'Concept Set'"
-                          [resourceName]=conceptSet.name
-                          [receiveDelete]=receiveDelete
-                          [closeFunction]="closeConfirmDelete">
-</app-confirm-delete-modal>
-
-<clr-modal [(clrModalOpen)]="removing" [clrModalSize]="'md'">
-  <div class="modal-title">
-    <div *ngIf="removing" class="modal-title-text">
-      Are you sure you want to remove {{selectedConceptsCount}} {{selectedConceptsCount > 1 ? 'concepts' : 'concept'}} from this set?
-    </div>
-  </div>
-  <div class="modal-footer">
-    <button type="button" class="btn btn-outline" (click)="removing=false">Cancel</button>
-    <button type="button" class="btn btn-primary confirm-remove-btn" (click)="removeConcepts()"
-            [clrLoading]="removeSubmitting" [disabled]="removeSubmitting">Remove Concepts</button>
-  </div>
-</clr-modal>
-<div *ngIf="!conceptSet" style="position: absolute; top: 50%; left: 50%;" class="spinner"></div>
