+++ ui/src/app/views/concept-set-list/component.tsx
@@ -13,11 +13,13 @@ import {reactStyles, ReactWrapperBase, withCurrentWorkspace} from 'app/utils/ind
 import {CreateConceptSetModal} from 'app/views/conceptset-create-modal/component';
 
 import {CardButton} from 'app/components/buttons';
+import {FadeBox} from 'app/components/containers';
 import {ClrIcon} from 'app/components/icons';
 import {TooltipTrigger} from 'app/components/popups';
 import {SpinnerOverlay} from 'app/components/spinners';
 import {convertToResources, ResourceType} from 'app/utils/resourceActions';
 import {WorkspaceData} from 'app/utils/workspace-data';
+import {ConceptNavigationBar} from 'app/views/concept-navigation-bar/component';
 import {ResourceCard} from 'app/views/resource-card/component';
 
 const styles = reactStyles({
@@ -78,37 +80,38 @@ export const ConceptSetsList = withCurrentWorkspace()(
 
     render() {
       const {conceptSetsLoading, conceptSetsList, conceptDomainList, createModalOpen} = this.state;
+      const {namespace, id} = this.props.workspace;
       const accessLevel = this.props.workspace.accessLevel as unknown as WorkspaceAccessLevel;
       const writePermission = accessLevel === WorkspaceAccessLevel.OWNER ||
           accessLevel === WorkspaceAccessLevel.WRITER;
 
-      return <React.Fragment>
-            <div style={styles.pageArea}>
-              <TooltipTrigger content={!writePermission &&
-              `Write permission required to create cohorts`} side='top'>
-                <CardButton type='small'
-                            onClick={() => this.setState({createModalOpen: true})}
-                            disabled={!writePermission}>
-                  Create a <br/>Concept Set
-                  <ClrIcon shape='plus-circle' size={21} style={{marginTop: 5}}/>
-                </CardButton>
-              </TooltipTrigger>
-              <div style={styles.resourceCardArea}>
-                {conceptSetsList && conceptSetsList.map((conceptSet: RecentResource) => {
-                  return <ResourceCard resourceCard={conceptSet} key={conceptSet.conceptSet.name}
-                                       onUpdate={() => this.loadConceptSets()}>
-                  </ResourceCard>;
-                })}
-              </div>
-              {conceptSetsLoading && <SpinnerOverlay />}
-            </div>
-
+      return <FadeBox style={{margin: 'auto', marginTop: '1rem', width: '95.7%'}}>
+        <ConceptNavigationBar wsId={id} ns={namespace} showConcepts={false}/>
+        <div style={styles.pageArea}>
+          <TooltipTrigger content={!writePermission &&
+          `Write permission required to create cohorts`} side='top'>
+            <CardButton type='small'
+                        onClick={() => this.setState({createModalOpen: true})}
+                        disabled={!writePermission}>
+              Create a <br/>Concept Set
+              <ClrIcon shape='plus-circle' size={21} style={{marginTop: 5}}/>
+            </CardButton>
+          </TooltipTrigger>
+          <div style={styles.resourceCardArea}>
+            {conceptSetsList && conceptSetsList.map((conceptSet: RecentResource) => {
+              return <ResourceCard resourceCard={conceptSet} key={conceptSet.conceptSet.name}
+                                   onUpdate={() => this.loadConceptSets()}>
+              </ResourceCard>;
+            })}
+          </div>
+          {conceptSetsLoading && <SpinnerOverlay />}
+        </div>
         {createModalOpen &&
         <CreateConceptSetModal onCreate={() => this.loadConceptSets()}
                                onClose={() => this.setState({createModalOpen: false})}
                                conceptDomainList={conceptDomainList}
                                existingConceptSets={conceptSetsList}/>}
-          </React.Fragment>;
+      </FadeBox>;
     }
 
   });
